adding more testing, making the __init__ files blank
