--- cg_openmm/__init__.py
@@ -1,7 +1,0 @@
-from . import build
-from . import cg_model
-from . import ensembles
-from . import parameters
-from . import simulation
-from . import thermo
-from . import utilities
