+++ cg_openmm/cg_model/__init__.py
@@ -1,1 +1,1 @@
-from .cgmodel import *
+
