--- cg_openmm/ensembles/__init__.py
@@ -1,1 +1,0 @@
-from .ens_build import *
