--- cg_openmm/parameters/__init__.py
@@ -1,3 +1,0 @@
-from .secondary_structure import *
-from .reweight import *
-from .optimize import *
