+++ cg_openmm/thermo/__init__.py
@@ -1,1 +1,1 @@
-from .calc import *
+
